combining 2 versions of root index to make it easier to add more projects
