+++ template/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@react-native-community/netinfo": "9.3.7",
-    "@sleeperhq/mini-core": "1.1.0",
+    "@sleeperhq/mini-core": "1.2.0",
     "axios": "0.15.3",
     "lodash": "4.17.21",
     "react": "17.0.2",
