+++ template/src/Fetch/index.tsx
@@ -1,81 +1,345 @@
-import React, {useCallback, useState} from 'react';
-import {
-  AppRegistry,
-  ActivityIndicator,
-  Text,
-  View,
-  StyleSheet,
-} from 'react-native';
-import {DevServer, Types} from '@sleeperhq/mini-core';
+import React, {useEffect, useState} from 'react';
+import * as RN from 'react-native';
+import {Types, Sleeper} from '@sleeperhq/mini-core';
+import {RostersMap} from '@sleeperhq/mini-core/declarations/types';
+import RosterOwners from './RosterOwners';
 
-import Fetch from './Fetch';
-import config from '../../app.json';
+type OwnProps = {
+  context: Types.Context;
+};
 
-DevServer.init(config);
+type Mode = {
+  name: string;
+  render: (props: OwnProps) => JSX.Element | null;
+};
 
-const Template = () => {
-  const [context, setContext] = useState<Types.Context>({} as Types.Context);
-  const [connected, setConnected] = useState<boolean>(false);
-  const [, updateState] = React.useState<any>();
-  const forceUpdate = React.useCallback(() => updateState({}), []);
+const Fetch = (props: OwnProps) => {
+  const {leaguesMap, rostersInLeagueMap, playersInSportMap} = props.context;
 
-  const _onContextChanged = useCallback((data: Types.Context) => {
-    setContext(data);
-  }, []);
+  const [selectedLeague, setSelectedLeague] = useState<string>();
+  const [selectedMode, setSelectedMode] = useState<Mode>();
+  const [selectedSport, setSelectedSport] = useState<string>();
+  const [selectedRosterMap, setSelectedRosterMap] = useState<RostersMap>();
 
-  const _onContextUpdated = useCallback(
-    (data: any) => {
-      setContext(existing => {
-        for (const key in data) {
-          existing[key] = data[key];
-        }
-        return existing;
-      });
-      forceUpdate();
-    },
-    [forceUpdate],
-  );
+  useEffect(() => {
+    if (selectedLeague) {
+      const sport = leaguesMap[selectedLeague]?.sport ?? '';
+      setSelectedSport(sport);
+    }
+  }, [selectedLeague, leaguesMap]);
 
-  const _onConnected = useCallback((value: boolean) => {
-    setConnected(value);
-  }, []);
+  useEffect(() => {
+    if (selectedLeague) {
+      const rosterMap = rostersInLeagueMap[selectedLeague];
+      setSelectedRosterMap(rosterMap);
+    }
+  }, [selectedLeague, rostersInLeagueMap]);
 
-  const _renderWaitingForConnection = () => {
+  useEffect(() => {
+    setSelectedMode(undefined);
+  }, [selectedLeague]);
+
+  const renderLeagueList = (props: OwnProps) => {
+    const {userLeagueList, leaguesMap} = props.context;
     return (
-      <View style={styles.loadingContainer}>
-        <Text style={styles.loadingText}>Waiting for connection...</Text>
-        <ActivityIndicator size={50} />
-      </View>
+      <RN.View style={styles.itemContainer}>
+        <Sleeper.Text style={styles.header}>Pick a League:</Sleeper.Text>
+        <RN.FlatList
+          style={styles.scroll}
+          data={userLeagueList}
+          renderItem={({item}) => (
+            <Sleeper.Button
+              text={leaguesMap[item].name || item}
+              onPress={() => setSelectedLeague(item)}
+            />
+          )}
+        />
+      </RN.View>
     );
   };
 
-  return (
-    <View style={styles.container}>
-      <DevServer
-        onContextChanged={_onContextChanged}
-        onContextUpdated={_onContextUpdated}
-        onConnected={_onConnected}
+  const renderModeList = (props: OwnProps) => {
+    const {leaguesMap} = props.context;
+
+    return (
+      <RN.View style={styles.itemContainer}>
+        <Sleeper.Text style={styles.header}>
+          Select Mode ({!!selectedLeague && leaguesMap[selectedLeague].name}):
+        </Sleeper.Text>
+        <RN.FlatList
+          style={styles.scroll}
+          data={modes}
+          renderItem={({item}) => (
+            <Sleeper.Button
+              text={item.name}
+              onPress={() => setSelectedMode(item)}
+            />
+          )}
+        />
+      </RN.View>
+    );
+  };
+
+  const renderModeHeader = () => {
+    if (!selectedLeague || !selectedMode) {
+      return null;
+    }
+
+    return (
+      <Sleeper.Text style={styles.header}>{selectedMode.name}</Sleeper.Text>
+    );
+  };
+
+  const renderMode = (props: OwnProps) => {
+    if (!selectedLeague) {
+      return null;
+    }
+
+    return (
+      <RN.View>
+        {renderModeList(props)}
+        {!!selectedMode && (
+          <RN.View style={styles.itemContainer}>
+            {renderModeHeader()}
+            {!!selectedMode && selectedMode.render(props)}
+          </RN.View>
+        )}
+      </RN.View>
+    );
+  };
+
+  const renderRosters = (props: OwnProps) => {
+    const {userMap} = props.context;
+
+    if (!selectedRosterMap || !selectedLeague) {
+      return null;
+    }
+
+    return <RosterOwners rostersMap={selectedRosterMap} userMap={userMap} />;
+  };
+
+  const renderUsers = (props: OwnProps) => {
+    const {usersInLeagueMap, userMap} = props.context;
+
+    const leagueUserIdList =
+      !!selectedLeague &&
+      !!usersInLeagueMap[selectedLeague] &&
+      Object.keys(usersInLeagueMap[selectedLeague]);
+
+    if (!leagueUserIdList) {
+      return null;
+    }
+
+    return (
+      <RN.FlatList
+        style={styles.scroll}
+        data={leagueUserIdList}
+        renderItem={({item}) => (
+          <Sleeper.Text style={styles.text}>
+            {userMap[item].display_name}
+          </Sleeper.Text>
+        )}
       />
-      {connected && <Fetch context={context} />}
-      {!connected && _renderWaitingForConnection()}
-    </View>
+    );
+  };
+
+  const renderPlayoffs = (props: OwnProps) => {
+    const {playoffsInLeagueMap} = props.context;
+
+    if (!selectedLeague || !playoffsInLeagueMap[selectedLeague]) {
+      return null;
+    }
+
+    return (
+      <RN.FlatList
+        style={styles.scroll}
+        data={playoffsInLeagueMap[selectedLeague].bracket}
+        renderItem={({item}) => (
+          <Sleeper.Text style={styles.text}>
+            {selectedRosterMap[item.t1]?.owner_id ?? 'No Owner'} vs{' '}
+            {selectedRosterMap[item.t2]?.owner_id ?? 'No Owner'}
+          </Sleeper.Text>
+        )}
+      />
+    );
+  };
+
+  const renderSportInfo = (props: OwnProps) => {
+    const {sportInfoMap} = props.context;
+
+    if (!selectedLeague || !selectedSport) {
+      return null;
+    }
+
+    return (
+      <RN.View>
+        <Sleeper.Text style={styles.header}>
+          Sport: {selectedSport}
+        </Sleeper.Text>
+        <Sleeper.Text style={styles.header}>
+          League season: {sportInfoMap[selectedSport]?.league_season ?? 'No season'}:
+        </Sleeper.Text>
+      </RN.View>
+    );
+  };
+
+  const renderTransactions = (props: OwnProps) => {
+    const {transactionsInLeagueMap, transactionsMap, userMap} = props.context;
+
+    if (
+      !selectedLeague ||
+      !transactionsInLeagueMap[selectedLeague] ||
+      !transactionsMap ||
+      !userMap
+    ) {
+      return null;
+    }
+
+    return (
+      <RN.FlatList
+        style={styles.scroll}
+        data={transactionsInLeagueMap[selectedLeague]}
+        renderItem={({item}) => {
+          const creator = transactionsMap[item]?.creator;
+          if (!creator) {
+            return null;
+          }
+
+          return (
+            <Sleeper.Text style={styles.text}>
+              {userMap[creator]?.display_name}
+            </Sleeper.Text>
+          );
+        }}
+      />
+    );
+  };
+
+  const renderDrafts = (props: OwnProps) => {
+    const { draftsInLeagueMap, draftPicksInDraftMap} = props.context;
+    if (!selectedLeague || !draftsInLeagueMap[selectedLeague]) {
+      return null;
+    }
+
+    return (
+      <RN.FlatList
+        style={styles.scroll}
+        data={draftsInLeagueMap[selectedLeague]}
+        renderItem={({item}) => {
+          if (
+            !item.draft_id ||
+            !selectedSport ||
+            !draftPicksInDraftMap[item.draft_id]
+          ) {
+            return null;
+          }
+
+          const topPickId = draftPicksInDraftMap[item.draft_id][0]?.player_id;
+          if (!topPickId) {
+            return null;
+          }
+          const topPlayer = playersInSportMap[selectedSport][topPickId]
+
+          return (
+            <Sleeper.Text style={styles.text}>
+              Type: {item.type} - Top Pick:{' '}
+              {topPlayer.first_name + ' ' + topPlayer.last_name}
+            </Sleeper.Text>
+          );
+        }}
+      />
+    );
+  };
+
+  const renderDraftPickTrades = (props: OwnProps) => {
+    const {draftPickTradesInLeagueMap} = props.context;
+
+    if (!selectedLeague || !draftPickTradesInLeagueMap[selectedLeague]) {
+      return null;
+    }
+
+    return (
+      <RN.FlatList
+        style={styles.scroll}
+        data={draftPickTradesInLeagueMap[selectedLeague]}
+        renderItem={({item}) => {
+          if (!item.roster_id || !item.owner_id) {
+            return null;
+          }
+          return (
+            <Sleeper.Text style={styles.text}>
+              {selectedRosterMap[item.roster_id]?.owner_id} to{' '}
+              {selectedRosterMap[item.owner_id]?.owner_id}
+            </Sleeper.Text>
+          );
+        }}
+      />
+    );
+  };
+
+  const modes: Mode[] = [
+    {name: 'Rosters', render: renderRosters},
+    {name: 'Users', render: renderUsers},
+    {name: 'Playoffs', render: renderPlayoffs},
+    {name: 'Sport Info', render: renderSportInfo},
+    {name: 'Transactions', render: renderTransactions},
+    {name: 'Drafts', render: renderDrafts},
+    {name: 'Draft Pick Trades', render: renderDraftPickTrades},
+  ];
+
+  return (
+    <RN.View style={styles.container}>
+      {renderLeagueList(props)}
+      {renderMode(props)}
+    </RN.View>
   );
 };
 
-const styles = StyleSheet.create({
+const styles = RN.StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#18202f',
+    justifyContent: 'center',
+    alignItems: 'center',
   },
-  loadingText: {
-    color: 'white',
+  userAvatar: {
+    width: 50,
+    height: 50,
+  },
+  leagueAvatar: {
+    width: 25,
+    height: 25,
+  },
+  text: {
     fontSize: 20,
   },
-  loadingContainer: {
-    flex: 1,
+  header: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  horizontal: {
+    flexDirection: 'row',
+  },
+  itemContainer: {
+    backgroundColor: 'grey',
     alignItems: 'center',
     justifyContent: 'center',
+    borderColor: 'white',
+    borderWidth: 2,
+    borderRadius: 10,
+    padding: 20,
+    margin: 5,
+  },
+  jersey: {
+    width: 50,
+    height: 50,
+  },
+  scroll: {
+    height: 150,
+    flexGrow: 0,
+  },
+  row: {
+    padding: 2,
   },
 });
 
-AppRegistry.registerComponent(config.name, () => Template);
+export default Fetch;
